% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{argument_handlers}
\alias{argument_handlers}
\alias{handle_arg_call}
\title{Argument Handlers}
\usage{
handle_arg_call(call = NULL, env = NULL)
}
\arguments{
\item{call}{\verb{[language, NULL]} (optional, default \code{NULL})
\itemize{
\item \code{language}: an R language object such as one produced by \link{match.call} or
\code{quote}; this call will be reported in an error
\item \code{NULL}: the call is attempted to be inferred
}}

\item{env}{\verb{[NULL, environment]} (no default)
\itemize{
\item \code{environment}: use this environment
\item \code{NULL}: use \code{parent.frame(2)}
}

When \code{call} is \code{NULL}, it will be guessed to be
\code{eval(quote(match.call()), envir = env)}.}
}
\description{
Functions which handle arguments of check functions.
}
