% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion_types.R
\name{assertion_types}
\alias{assertion_types}
\alias{dev_assertion_types}
\title{Assertion Types}
\usage{
assertion_types()

dev_assertion_types()
}
\value{
\verb{[dbc::assertion_types]} returns vector
c("general", "input", "user_input", "prod_input", "dev_input",
"prod_output", "dev_output", "prod_interim", "dev_interim", "none"
)

\verb{[dbc::dev_assertion_types]} returns vector
c("dev_input", "dev_output", "dev_interim")
}
\description{
Tools to process assertion types.
}
