% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_theta_names}
\alias{build_theta_names}
\title{Build human-readable names for the θ vector}
\usage{
build_theta_names(parsed, dm_mats)
}
\description{
Build human-readable names for the θ vector
}
\keyword{internal}
