% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/manually_written_caller_environment_check_funs.R
\name{caller_environment_checks}
\alias{caller_environment_checks}
\alias{report_function_caller_environment_is_not_global_environment}
\alias{report_function_caller_environment_is_global_environment}
\alias{assert_function_caller_environment_is_not_global_environment}
\alias{assert_function_caller_environment_is_global_environment}
\alias{test_function_caller_environment_is_not_global_environment}
\alias{test_function_caller_environment_is_global_environment}
\title{Caller Environment Checks}
\usage{
report_function_caller_environment_is_not_global_environment(
  x = NULL,
  x_nm = NULL,
  call = NULL
)

report_function_caller_environment_is_global_environment(
  x = NULL,
  x_nm = NULL,
  call = NULL
)

assert_function_caller_environment_is_not_global_environment(
  x = NULL,
  x_nm = NULL,
  call = NULL
)

assert_function_caller_environment_is_global_environment(
  x = NULL,
  x_nm = NULL,
  call = NULL
)

test_function_caller_environment_is_not_global_environment(
  x = NULL,
  x_nm = NULL,
  call = NULL
)

test_function_caller_environment_is_global_environment(
  x = NULL,
  x_nm = NULL,
  call = NULL
)
}
\arguments{
\item{x}{\verb{[NULL, environment]} (default \code{NULL})
\itemize{
\item \code{environment}: this environment is compared against \verb{[globalenv]} output
\item \code{NULL}: \code{x} is taken to be \code{parent.frame(2L)} in the check function,
which should be the caller environment of the function you are trying
to check; See \strong{Examples}
}}

\item{x_nm}{\verb{[NULL, character]} (default \code{NULL})
\itemize{
\item \code{NULL}: The name of the function you are checking is guessed via
\code{eval(quote(match.call()), parent.frame(1L))}
\item \code{character}: This should be the name of the function you are checking.
}}

\item{call}{\verb{[language, NULL]} (optional, default \code{NULL})
\itemize{
\item \code{language}: an R language object such as one produced by \link{match.call} or
\code{quote}; this call will be reported in an error
\item \code{NULL}: the call is attempted to be inferred
}}
}
\description{
Functions to check caller environment of a function.
}
\examples{
# for use in other functions only
my_fun_ <- function() {
  dbc::report_function_caller_environment_is_not_global_environment()
}
result <- my_fun_()
stopifnot(is.data.frame(result), "message" \%in\% names(result),
          grepl("my_fun_", result[["message"]]))
my_fun_ <- function() {
  dbc::test_function_caller_environment_is_not_global_environment()
}
result <- my_fun_()
stopifnot(identical(result, FALSE))

# for use by the user only
my_fun <- function() {
  dbc::test_function_caller_environment_is_global_environment()
}
result <- my_fun()
stopifnot(identical(result, TRUE))

# for use by the user only
my_fun <- function() {
  dbc::assert_function_caller_environment_is_global_environment()
  my_fun_()
}
result <- my_fun()
stopifnot(identical(result, TRUE))

# this would be a mistake
my_other_fun <- function() {
  my_fun()
}
result <- tryCatch(my_other_fun(), error = function(e) e)
stopifnot(
  is.character(result[["message"]]),
  grepl("global", result[["message"]])
)

}
