% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_categorical.R
\name{clean_categorical}
\alias{clean_categorical}
\title{Clean categorical variables within a dataset based on a dictionary of
value-replacement pairs}
\usage{
clean_categorical(
  x,
  dict_allowed,
  dict_clean = NULL,
  vars_id = NULL,
  col_allowed_var = "variable",
  col_allowed_value = "value",
  fn = std_text,
  na = ".na"
)
}
\arguments{
\item{x}{A data frame with one or more columns to clean}

\item{dict_allowed}{Dictionary of allowed values for each variable of
interest. Must include columns for "variable" and "value" (the names of
which can be modified with args \code{col_allowed_var} and \code{col_allowed_value}).}

\item{dict_clean}{Optional dictionary of value-replacement pairs (e.g.
produced by \code{\link{check_categorical}}). Must include columns "variable",
"value", "replacement", and, if specified as an argument, all columns
specified by \code{vars_id}.

If no dictionary is provided, will simply standardize columns to match
allowed values specified in \code{dict_allowed}.}

\item{vars_id}{Optional vector of one or more ID columns within \code{x} on which
corrections should be conditional.

If not specified the cleaning dictionary contains one entry for each unique
combination of variable and non-valid value. If specified the cleaning
dictionary contains one entry for each unique combination of variable,
non-valid value, and ID variable.}

\item{col_allowed_var}{Name of column in \code{dict_allowed} giving variable name
(defaults to "variable")}

\item{col_allowed_value}{Name of column in \code{dict_allowed} giving allowed
values (defaults to "value")}

\item{fn}{Function to standardize raw values in both the dataset and
dictionary prior to comparing, to account for minor variation in character
case, spacing, punctuation, etc. Defaults to \code{\link{std_text}}. To omit the
standardization step can use e.g. \code{as.character} or an identity function
\code{function(x) x}.}

\item{na}{Keyword to use within column "replacement" for values that should
be converted to \code{NA}. Defaults to ".na". The keyword is used to distinguish
between "replacement" values that are missing because they have yet to be
manually verified, and values that have been verified and really should be
converted to \code{NA}.}
}
\value{
The original data frame \code{x} but with cleaned versions of the categorical
variables specified in argument \code{dict_allowed}
}
\description{
Applies a dictionary of value-replacement pairs to clean and standardize
values of categorical variables. Includes options for text standardization to
standardize minor differences in character case, spacing, and punctuation.
}
\examples{
# load example dataset, dictionary of allowed categorical values, and
# cleaning dictionary
data(ll1)
data(dict_categ1)
data(clean_categ1)

# dictionary-based corrections to categorical vars
clean_categorical(
  ll1,
  dict_allowed = dict_categ1,
  dict_clean = clean_categ1
)

# require exact matching, including character case
clean_categorical(
  ll1,
  dict_allowed = dict_categ1,
  dict_clean = clean_categ1,
  fn = identity
)

# apply standardization to dict_allowed but no additional dict-based cleaning
clean_categorical(
  ll1,
  dict_allowed = dict_categ1
)

}
