% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clean_num1}
\alias{clean_num1}
\title{A cleaning dictionary for numeric variables in example dataset \code{ll1}}
\format{
A data.frame with 4 rows and 4 variables:
\describe{
\item{variable}{Column name within dataset}
\item{value}{Non-valid numeric value}
\item{replacement}{Replacement value for given non-valid value}
\item{new}{Logical indicating whether the dictionary entry is new}
}
}
\usage{
clean_num1
}
\description{
A cleaning dictionary for numeric variables in example dataset \code{ll1}
}
\keyword{datasets}
