% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.dbc}
\alias{coef.dbc}
\title{Extract coefficients from a dbc object}
\usage{
\method{coef}{dbc}(object, ...)
}
\arguments{
\item{object}{A \code{dbc} object.}

\item{...}{Currently unused.}
}
\description{
Extract coefficients from a dbc object
}
\examples{
set.seed(1)
dat <- DGP(N = 100, N_T = 4, rho1 = 0.3, rho2 = 0.2, tau = 0.5)
fit <- dbc(
    outcome_fml   = y ~ lag_y + D,
    treatment_fml = D ~ lag_y,
    lag_y = "lag_y", treatment = "D",
    panel_id = "panel_id", time_id = "time",
    data = dat
)
coef(fit)
}
