% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dev_mode}
\alias{dev_mode}
\alias{set_dev_mode}
\alias{get_dev_mode}
\title{Development Mode}
\usage{
set_dev_mode(value)

get_dev_mode()
}
\arguments{
\item{value}{\verb{[logical]} (mandatory, no default)

if set to \code{TRUE}, "dev" assertions will be evaluated; else they won't be}
}
\description{
Set and get development mode for \strong{dbc} functions.
}
