% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict_categ1}
\alias{dict_categ1}
\title{A dictionary of allowed values for categorical variables in example dataset
\code{ll1}}
\format{
A data.frame with 13 rows and 2 variables:
\describe{
\item{variable}{Column name within dataset}
\item{value}{Allowed categorical values for given column}
}
}
\usage{
dict_categ1
}
\description{
A dictionary of allowed values for categorical variables in example dataset
\code{ll1}
}
\keyword{datasets}
