% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_data}
\alias{error_data}
\alias{get_error_data}
\alias{get_newest_error_data}
\title{Error Data}
\usage{
get_error_data(n = 1L)

get_newest_error_data()
}
\arguments{
\item{n}{\verb{[integer]} (optional, default \code{1L})

position of error in list of errors in chronological order recorded during
this R session; \code{n = 1L} is the position of the newest error}
}
\value{
Both \code{get_error_data} and \code{get_newest_error_data} return a list containing
elements
\itemize{
\item \code{msg}: \code{character}; error message
\item \code{call}: \code{call}; call where error was raised
\item \code{sys.calls}: output of \verb{[base::sys.calls]} in the context of where the
error was generated
}
}
\description{
Retrieve data about previous errors.
}
