% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{function_calls}
\alias{function_calls}
\alias{get_current_call}
\alias{get_parent_call}
\alias{get_nth_call}
\title{Function Calls}
\usage{
get_current_call()

get_parent_call()

get_nth_call(n = 1L)
}
\arguments{
\item{n}{\verb{[integer]} (default \code{1L})

Integer larger than zero. \code{1L} gives the call of the function which calls
\code{get_nth_call}.}
}
\description{
Retrieve upsteam function calls.
}
\section{Functions}{

\itemize{
\item \code{get_current_call}: when called inside another function, gives the call
of that function. See \strong{Examples}.
}


\itemize{
\item \code{get_parent_call}: when called inside another function, gives the call
of the function calling that one. See \strong{Examples}.
}


\itemize{
\item \code{get_nth_call}: when called inside another function, gives the \code{n}'th
parent call of that function.
}
}

\examples{
tf1 <- function(tf1_arg = 1) {
  get_current_call()
}
stopifnot(
  identical(tf1(), quote(tf1()))
)
tf2 <- function(tf2_arg = 2) {
  get_parent_call()
}
tf3 <- function(tf3_arg = 3) {
  tf2()
}
stopifnot(
  identical(tf3(), quote(tf3()))
)
tf4 <- function(tf4_arg = 4) {
  get_parent_call()
}
stopifnot(
  inherits(tryCatch(get_current_call(), error = function(e) e), "error"),
  inherits(tryCatch(get_parent_call(), error = function(e) e), "error"),
  inherits(tryCatch(tf5(), error = function(e) e), "error")
)
}
