% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_biased_ols}
\alias{get_biased_ols}
\title{Get biased OLS estimates as starting values for GMM}
\usage{
get_biased_ols(dm_mats, parsed)
}
\description{
Get biased OLS estimates as starting values for GMM
}
\keyword{internal}
