% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ll1}
\alias{ll1}
\title{An example messy dataset to clean}
\format{
A data.frame with 7 rows and 10 variables:
\describe{
\item{id}{Patient identifier}
\item{age}{Age value}
\item{age_unit}{Age units}
\item{sex}{Patient sex}
\item{status}{Patient status}
\item{contacts}{Number of epidemiological contacts}
\item{date_onset}{Date of symptom onset}
\item{date_admit}{Date of admission to hospital}
\item{date_exit}{Date of exit from hospital}
\item{exit_status}{Patient outcome status}
}
}
\usage{
ll1
}
\description{
An example messy dataset to clean
}
\keyword{datasets}
