% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nobs.dbc}
\alias{nobs.dbc}
\title{Number of observations used in a dbc model}
\usage{
\method{nobs}{dbc}(object, ...)
}
\arguments{
\item{object}{A \code{dbc} object.}

\item{...}{Currently unused.}
}
\description{
Returns the total number of panel observations (N x T).
}
\examples{
set.seed(1)
dat <- DGP(N = 100, N_T = 4, rho1 = 0.3, rho2 = 0.2, tau = 0.5)
fit <- dbc(
    outcome_fml   = y ~ lag_y + D,
    treatment_fml = D ~ lag_y,
    lag_y = "lag_y", treatment = "D",
    panel_id = "panel_id", time_id = "time",
    data = dat
)
nobs(fit)
}
