% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dates.R
\name{parse_dates}
\alias{parse_dates}
\title{Parse dates}
\usage{
parse_dates(
  x,
  convert_excel = TRUE,
  as_date = TRUE,
  orders = c("Ymd", "dmY", "dmy", "mdY", "Ymd HMS")
)
}
\arguments{
\item{x}{A character or numeric vector of dates}

\item{convert_excel}{Logical indicating whether to convert Excel-encoded date
values (e.g. "42370") into dates, using \link[janitor:excel_numeric_to_date]{janitor::excel_numeric_to_date}}

\item{as_date}{Logical indicating whether to convert to Date class using
\link[lubridate:as_date]{lubridate::as_date}, which will quietly force any non-parseable date value
to NA. Defaults to TRUE.}

\item{orders}{a character vector of date-time formats. Each order string is
a series of formatting characters as listed in \code{\link[base:strptime]{base::strptime()}} but
might not include the \code{"\%"} prefix. For example, "ymd" will match all the
possible dates in year, month, day order. Formatting orders might include
arbitrary separators. These are discarded. See details for the implemented
formats. If multiple order strings are supplied, they are applied in turn
for \code{parse_date_time2()} and \code{fast_strptime()}. For \code{parse_date_time()}
the order of applied formats is determined by \code{select_formats} parameter.}
}
\value{
If argument \code{as_date} is \code{TRUE}, a vector of class "Date", with values that
cannot be converted to valid dates returned as \verb{<NA>}. Else a vector of class
"character" with values that cannot be converted to valid dates retained as
is.
}
\description{
Parse dates
}
\examples{
x <- c("44087", "12//02/2019", "2020_05_14", "2021-01-30 14:00:04")
parse_dates(x)

}
