% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.dbc}
\alias{print.summary.dbc}
\title{Print a summary.dbc object}
\usage{
\method{print}{summary.dbc}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{summary.dbc} object.}

\item{digits}{Number of significant digits.}

\item{...}{Further arguments passed to \code{\link[=printCoefmat]{printCoefmat()}}.}
}
\description{
Print a summary.dbc object
}
\examples{
set.seed(1)
dat <- DGP(N = 100, N_T = 4, rho1 = 0.3, rho2 = 0.2, tau = 0.5)
fit <- dbc(
    outcome_fml   = y ~ lag_y + D,
    treatment_fml = D ~ lag_y,
    lag_y = "lag_y", treatment = "D",
    panel_id = "panel_id", time_id = "time",
    data = dat
)
print(summary(fit))
}
