% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manually_written_report_funs.R
\name{report_is_report_df}
\alias{report_is_report_df}
\title{\code{report_df} Report}
\usage{
report_is_report_df(x, x_nm = NULL, call = NULL)
}
\arguments{
\item{x}{R object to be tested (mandatory, no default)}

\item{x_nm}{\verb{[NULL, character]} (optional, default \code{NULL})

the name of the object \code{x} to mention in possible error message.
\itemize{
\item \code{NULL}: taken as \code{deparse(substitute(x))}
\item \code{character}: the name as a string
}}

\item{call}{\verb{[language, NULL]} (optional, default \code{NULL})
\itemize{
\item \code{language}: an R language object such as one produced by \link{match.call} or
\code{quote}; this call will be reported in an error
\item \code{NULL}: the call is attempted to be inferred
}}
}
\description{
Report whether an object is a valid report data.frame as produced by e.g.
\verb{[dbc::expressions_to_report]}.
}
