% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{run_dbc_gmm}
\alias{run_dbc_gmm}
\title{Run the DBC GMM estimation}
\usage{
run_dbc_gmm(dm_mats, parsed, N, N_T, start_vals, reltol = 1e-12, maxit = 10000)
}
\description{
Run the DBC GMM estimation
}
\keyword{internal}
