% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_text.R
\name{std_text}
\alias{std_text}
\title{Standardize text prior to matching to account for minor variation in
character case, spacing, punctuation, or use of accents}
\usage{
std_text(x)
}
\arguments{
\item{x}{A vector of strings}
}
\value{
The standardized version of \code{x}
}
\description{
Implements the following transformations:
\enumerate{
\item standardize case (\code{base::tolower})
\item remove diacritic/accent characters (\code{stringi::stri_trans_general})
\item remove sequences of space or punctuation characters at start or end of string
\item replace repeated whitespace characters with a single space
}
}
\examples{
std_text(c("CONFIRMED", "Conf.", "confirmed"))
std_text(c("R\u00e9publique d\u00e9mocratique du  Congo", "Nigeria_"))
}
