% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utilities}
\alias{utilities}
\alias{handle_x_nm_arg}
\title{Utilities}
\usage{
handle_x_nm_arg(x_nm)
}
\arguments{
\item{x_nm}{\verb{[NULL, character]} (mandatory, no default)
\itemize{
\item \code{character}: return input as-is
\item \code{NULL}: it is assumed that this function is called within another function,
and essentially \code{deparse(substitute(x))} is returned
}}
}
\value{
\itemize{
\item \code{handle_x_nm_arg}: always returns a character vector of length 1
}
}
\description{
Utility functions, intended primarily for internal use.
}
\details{
\itemize{
\item \code{handle_x_nm_arg} is used internally in \code{report_} and \code{assert_} functions
to guess the name of the object passed to argument \code{x} when it is not
supplied explicitly.
}
}
