% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate inputs to dbc()}
\usage{
validate_inputs(
  outcome_fml,
  treatment_fml,
  lag_y,
  lag_d,
  treatment,
  panel_id,
  time_id,
  data
)
}
\description{
Validate inputs to dbc()
}
\keyword{internal}
