% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{argument_handlers}
\alias{argument_handlers}
\alias{handle_arg_call}
\alias{handle_arg_x_nm}
\title{Argument Handlers}
\usage{
handle_arg_call(call = NULL, env = NULL)

handle_arg_x_nm(x_nm)
}
\arguments{
\item{call}{\verb{[language, NULL]} (optional, default \code{NULL})
\itemize{
\item \code{language}: an R language object such as one produced by \link{match.call} or
\code{quote}; this call will be reported in an error
\item \code{NULL}: the call is attempted to be inferred
}}

\item{env}{\verb{[NULL, environment]} (DEPRECATED)

Deprecated in 0.3.21.}

\item{x_nm}{\verb{[NULL, character]} (optional, default \code{NULL})

the name of the object \code{x} to mention in possible error message.
\itemize{
\item \code{NULL}: taken as \code{deparse(substitute(x))}
\item \code{character}: the name as a string
}}
}
\value{
\itemize{
\item \code{dbc::handle_arg_call}: returns an R \code{language} object, or \code{NULL} upon
failure to guess the call
}

\itemize{
\item \code{dbc::handle_x_nm_arg}: always returns a character vector of length 1
}
}
\description{
Functions which handle arguments of check functions.
}
\section{Functions}{

\itemize{
\item \code{dbc::handle_arg_call} is used internally in other functions
to guess \code{call} which is to be reported if there is a problem
}


\itemize{
\item \code{dbc::handle_x_nm_arg} is used internally in other functions
to guess the name of the object passed to argument \code{x} when it is not
supplied explicitly. It uses object \code{x} in its caller environment
to guess \code{x_nm} if it is \code{NULL}.
}
}

